+++ src/main/java/com/vapeshop/controller/UpdateProfileServlet.java
@@ -2,7 +2,7 @@ package com.vapeshop.controller;
 
 
 import com.vapeshop.entity.User;
-import com.vapeshop.respository.UserRespository;
+import com.vapeshop.respository.user.UserRespository;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
@@ -17,13 +17,13 @@ import java.io.IOException;
 public class UpdateProfileServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.getRequestDispatcher("update-profile.jsp").forward(request,response);
+        request.getRequestDispatcher("update-profile.jsp").forward(request, response);
     }
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        HttpSession session=request.getSession();
-      User userOld= (User) session.getAttribute("user");
+        HttpSession session = request.getSession();
+        User userOld = (User) session.getAttribute("user");
         String id = userOld.getId();
         String fullname = request.getParameter("fullname");
         String phone = request.getParameter("phone");
