+++ src/main/java/com/vapeshop/controller/employee/chat/ChatListServlet.java
@@ -1,13 +1,12 @@
 package com.vapeshop.controller.employee.chat;
 
 import com.vapeshop.chat.ChatSession;
-import com.vapeshop.lab.ChatEndpoint;
+import com.vapeshop.chat.ChatEndpoint;
 import jakarta.servlet.*;
 import jakarta.servlet.http.*;
 import jakarta.servlet.annotation.*;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 @WebServlet(name = "ChatListServlet", value = "/chat-list")
