+++ src/main/java/com/vapeshop/controller/employee/service/TestServlet.java
@@ -1,0 +1,24 @@
+package com.vapeshop.controller.employee.service;
+
+import jakarta.servlet.*;
+import jakarta.servlet.http.*;
+import jakarta.servlet.annotation.*;
+
+import java.io.IOException;
+
+@WebServlet(name = "TestServlet", value = "/update-test")
+public class TestServlet extends HttpServlet {
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        int i=02;
+        System.out.println(i);
+        String s="kkk";
+        System.out.println(s);
+        response.sendRedirect("index.jsp");
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+
+    }
+}
