+++ src/main/java/com/vapeshop/controller/product/ShowProductDetails.java
@@ -2,7 +2,7 @@ package com.vapeshop.controller.product;
 
 import com.vapeshop.entity.Product;
 import com.vapeshop.entity.ProductType;
-import com.vapeshop.respository.ProductRepository;
+import com.vapeshop.respository.user.ProductRepository;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
