--- src/main/java/com/vapeshop/respository/HelloRespository.java
@@ -1,4 +1,0 @@
-package com.vapeshop.respository;
-
-public class HelloRespository {
-}
