+++ src/main/java/com/vapeshop/respository/user/SendMailRespository.java
@@ -1,4 +1,4 @@
-package com.vapeshop.respository;
+package com.vapeshop.respository.user;
 
 import javax.mail.*;
 import javax.mail.internet.InternetAddress;
@@ -7,7 +7,7 @@ import javax.mail.internet.MimeMessage;
 import javax.mail.internet.MimeMultipart;
 import java.io.IOException;
 import java.util.Properties;
-public class SendMail {
+public class SendMailRespository {
     public static boolean forgotPass(String emailTo, String otp) {
         String to = emailTo; // to address. It can be any like gmail, hotmail etc.
         final String from = "petshopfptv1@gmail.com"; // from address. As this is using Gmail SMTP.
